ocr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
paddlepaddle==2.5.2
paddleocr==2.7.3
PyMuPDF==1.23.14
Pillow==10.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
numpy==1.24.4